+++ cmd/cli/query.go
@@ -218,11 +218,11 @@ var (
 	}
 
 	blkByHeightCmd = &cobra.Command{
-		Use:   "block-by-height <height> ",
+		Use:   "block <height> ",
 		Short: "query a block at a height",
 		Args:  cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
-			writeToConsole(client.CertByHeight(uint64(argToInt(args[0]))))
+			writeToConsole(client.BlockByHeight(uint64(argToInt(args[0]))))
 		},
 	}