+++ store/error.go
@@ -64,3 +64,7 @@ func ErrCompactProof(err error) lib.ErrorI {
 func ErrInvalidKey() lib.ErrorI {
 	return lib.NewError(lib.CodeInvalidKey, lib.StorageModule, "found store key is invalid")
 }
+
+func ErrOddLengthBranch() lib.ErrorI {
+	return lib.NewError(lib.CodeOddLengthMerkleBranch, lib.StorageModule, "the merkle branch length is odd")
+}